初始化-优化代码规范

+++ src/main/java/com/example/emos/wx/controller/MeetingController.java
@@ -34,73 +34,74 @@ public class MeetingController {
 
     @PostMapping("/searchMyMeetingListByPage")
     @ApiOperation("查询会议列表分页数据")
-    public R searchMyMeetingListByPage(@Valid @RequestBody SearchMyMeetingListByPageForm form,@RequestHeader("token") String token){
-        int userId=jwtUtil.getUserId(token);
-        int page=form.getPage();
-        int length=form.getLength();
-        long start=(page-1)*length;
-        HashMap map=new HashMap();
-        map.put("userId",userId);
-        map.put("start",start);
-        map.put("length",length);
-        ArrayList list=meetingService.searchMyMeetingListByPage(map);
-        return R.ok().put("result",list);
+    public R searchMyMeetingListByPage(@Valid @RequestBody SearchMyMeetingListByPageForm form, @RequestHeader("token") String token) {
+        int userId = jwtUtil.getUserId(token);
+        int page = form.getPage();
+        int length = form.getLength();
+        long start = (long) (page - 1) * length;
+        HashMap map = new HashMap();
+        map.put("userId", userId);
+        map.put("start", start);
+        map.put("length", length);
+        ArrayList list = meetingService.searchMyMeetingListByPage(map);
+        return R.ok().put("result", list);
     }
 
     @PostMapping("/insertMeeting")
     @ApiOperation("添加会议")
-    @RequiresPermissions(value = {"ROOT", "MEETING:INSERT"},logical = Logical.OR)
-    public R insertMeeting(@Valid @RequestBody InsertMeetingForm form,@RequestHeader("token") String token){
-        if(form.getType()==2&&(form.getPlace()==null||form.getPlace().length()==0)){
+    @RequiresPermissions(value = {"ROOT", "MEETING:INSERT"}, logical = Logical.OR)
+    public R insertMeeting(@Valid @RequestBody InsertMeetingForm form, @RequestHeader("token") String token) {
+        if (form.getType() == 2 && (form.getPlace() == null || form.getPlace().length() == 0)) {
             throw new EmosException("线下会议地点不能为空");
         }
-        DateTime d1= DateUtil.parse(form.getDate()+" "+form.getStart()+":00");
-        DateTime d2= DateUtil.parse(form.getDate()+" "+form.getEnd()+":00");
-        if(d2.isBeforeOrEquals(d1)){
+        DateTime d1 = DateUtil.parse(form.getDate() + " " + form.getStart() + ":00");
+        DateTime d2 = DateUtil.parse(form.getDate() + " " + form.getEnd() + ":00");
+        if (d2.isBeforeOrEquals(d1)) {
             throw new EmosException("结束时间必须大于开始时间");
         }
-        if(!JSONUtil.isJsonArray(form.getMembers())){
+        if (!JSONUtil.isJsonArray(form.getMembers())) {
             throw new EmosException("members不是JSON数组");
         }
-        TbMeeting entity=new TbMeeting();
+        TbMeeting entity = new TbMeeting();
         entity.setUuid(UUID.randomUUID().toString(true));
         entity.setTitle(form.getTitle());
-        entity.setCreatorId((long)jwtUtil.getUserId(token));
+        entity.setCreatorId((long) jwtUtil.getUserId(token));
         entity.setDate(form.getDate());
         entity.setPlace(form.getPlace());
         entity.setStart(form.getStart() + ":00");
         entity.setEnd(form.getEnd() + ":00");
-        entity.setType((short)form.getType());
+        entity.setType((short) form.getType());
         entity.setMembers(form.getMembers());
         entity.setDesc(form.getDesc());
-        entity.setStatus((short)1);
+        entity.setStatus((short) 1);
         meetingService.insertMeeting(entity);
-        return R.ok().put("result","success");
+        return R.ok().put("result", "success");
     }
+
     @PostMapping("/searchMeetingById")
     @ApiOperation("根据ID查询会议")
     @RequiresPermissions(value = {"ROOT", "MEETING:SELECT"}, logical = Logical.OR)
-    public R searchMeetingById(@Valid @RequestBody SearchMeetingByIdFrom form){
-        HashMap map=meetingService.searchMeetingById(form.getId());
-        return R.ok().put("result",map);
+    public R searchMeetingById(@Valid @RequestBody SearchMeetingByIdFrom form) {
+        HashMap map = meetingService.searchMeetingById(form.getId());
+        return R.ok().put("result", map);
     }
 
     @PostMapping("/updateMeetingInfo")
     @ApiOperation("更新会议")
     @RequiresPermissions(value = {"ROOT", "MEETING:UPDATE"}, logical = Logical.OR)
-    public R updateMeetingInfo(@Valid @RequestBody UpdateMeetingInfoForm form){
-        if(form.getType()==2&&(form.getPlace()==null||form.getPlace().length()==0)){
+    public R updateMeetingInfo(@Valid @RequestBody UpdateMeetingInfoForm form) {
+        if (form.getType() == 2 && (form.getPlace() == null || form.getPlace().length() == 0)) {
             throw new EmosException("线下会议地点不能为空");
         }
-        DateTime d1= DateUtil.parse(form.getDate()+" "+form.getStart()+":00");
-        DateTime d2= DateUtil.parse(form.getDate()+" "+form.getEnd()+":00");
-        if(d2.isBeforeOrEquals(d1)){
+        DateTime d1 = DateUtil.parse(form.getDate() + " " + form.getStart() + ":00");
+        DateTime d2 = DateUtil.parse(form.getDate() + " " + form.getEnd() + ":00");
+        if (d2.isBeforeOrEquals(d1)) {
             throw new EmosException("结束时间必须大于开始时间");
         }
-        if(!JSONUtil.isJsonArray(form.getMembers())){
+        if (!JSONUtil.isJsonArray(form.getMembers())) {
             throw new EmosException("members不是JSON数组");
         }
-        HashMap param=new HashMap();
+        HashMap param = new HashMap();
         param.put("title", form.getTitle());
         param.put("date", form.getDate());
         param.put("place", form.getPlace());
@@ -113,43 +114,43 @@ public class MeetingController {
         param.put("instanceId", form.getInstanceId());
         param.put("status", 1);
         meetingService.updateMeetingInfo(param);
-        return R.ok().put("result","success");
+        return R.ok().put("result", "success");
     }
+
     @PostMapping("/deleteMeetingById")
     @ApiOperation("根据ID删除会议")
     @RequiresPermissions(value = {"ROOT", "MEETING:DELETE"}, logical = Logical.OR)
-    public R deleteMeetingById(@Valid @RequestBody DeleteMeetingByIdForm form){
+    public R deleteMeetingById(@Valid @RequestBody DeleteMeetingByIdForm form) {
         meetingService.deleteMeetingById(form.getId());
-        return R.ok().put("result","success");
+        return R.ok().put("result", "success");
     }
 
     @PostMapping("/recieveNotify")
     @ApiOperation("接收工作流通知")
-    public R recieveNotify(@Valid @RequestBody RecieveNotifyForm form){
-        if(form.getResult().equals("同意")){
-            log.debug(form.getUuid()+"的会议审批通过");
-        }
-        else{
-            log.debug(form.getUuid()+"的会议审批不通过");
+    public R recieveNotify(@Valid @RequestBody RecieveNotifyForm form) {
+        if (form.getResult().equals("同意")) {
+            log.debug(form.getUuid() + "的会议审批通过");
+        } else {
+            log.debug(form.getUuid() + "的会议审批不通过");
         }
         return R.ok();
     }
 
     @PostMapping("/searchRoomIdByUUID")
     @ApiOperation("查询会议房间RoomID")
-    public R searchRoomIdByUUID(@Valid @RequestBody SearchRoomIdByUUIDForm form){
-        long roomId=meetingService.searchRoomIdByUUID(form.getUuid());
-        return R.ok().put("result",roomId);
+    public R searchRoomIdByUUID(@Valid @RequestBody SearchRoomIdByUUIDForm form) {
+        long roomId = meetingService.searchRoomIdByUUID(form.getUuid());
+        return R.ok().put("result", roomId);
     }
 
     @PostMapping("/searchUserMeetingInMonth")
     @ApiOperation("查询某月用户的会议日期列表")
-    public R searchUserMeetingInMonth(@Valid @RequestBody SearchUserMeetingInMonthForm form,@RequestHeader("token") String token){
-        int userId=jwtUtil.getUserId(token);
-        HashMap param=new HashMap();
-        param.put("userId",userId);
-        param.put("express",form.getYear()+"/"+form.getMonth());
-        List<String> list=meetingService.searchUserMeetingInMonth(param);
-        return R.ok().put("result",list);
+    public R searchUserMeetingInMonth(@Valid @RequestBody SearchUserMeetingInMonthForm form, @RequestHeader("token") String token) {
+        int userId = jwtUtil.getUserId(token);
+        HashMap param = new HashMap();
+        param.put("userId", userId);
+        param.put("express", form.getYear() + "/" + form.getMonth());
+        List<String> list = meetingService.searchUserMeetingInMonth(param);
+        return R.ok().put("result", list);
     }
 }
